serve html content in unique link
